adição de bibliotecas ao main
Para compilar usando o gcc:
"gcc carta_atleta.c -o atleta -I./lib -L./lib -latleta"
